Implement simple references in Celia runtime

+++ runtime/main.c
@@ -1,8 +1,8 @@
 #include "Lumina.h"
 
-long long f_0main(long long _0arg);
+int64_t f_0main(int64_t _0arg);
 
 int main() {
-    printf("%lld\n", f_0main(0));
+    printf("%ld\n", f_0main(0));
     return 0;
 }
